Rename a subfolder
