+++ src/Functions/AI/LLMFunctionBase.h
@@ -2,9 +2,9 @@
 
 #include <Functions/IFunction.h>
 #include <Functions/FunctionFactory.h>
-#include <Functions/LLM/ILLMProvider.h>
-#include <Functions/LLM/LLMResultCache.h>
-#include <Functions/LLM/LLMQuotaTracker.h>
+#include <Functions/AI/ILLMProvider.h>
+#include <Functions/AI/LLMResultCache.h>
+#include <Functions/AI/LLMQuotaTracker.h>
 #include <Interpreters/Context.h>
 
 namespace DB
